+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "has-go" extension will be documented in this file.
 
 - Added `Stack Build`
 - Added `Stack Test`
+- Added `Compile Haskell File`
 - Updated `overrideGHCiArgs`
 
 ## [0.3.0]